Miscellaneous fixes - can't remember what exactly
I haven't worked on this for a while and can't remember exactly what I
changed here. It seems to be quite a lot of different things. Including
navigation + implementing question strength as a float. However it
does't run - probably because of differences in the way js defines
objects and swift syntax